refactor: extract common nostr identifier parsing logic
- Added shared utility functions in nostrIdentifiers.ts:
  - normalizeNostrIdentifier: removes nostr: prefix and normalizes
  - parseEventIdentifier: handles nevent, note, and hex event IDs for /e/[id]
  - parseProfileIdentifier: handles nprofile, npub for /p/[id]
  - isValidEventIdentifier: checks if identifier is valid nevent/note
  - isValidNpub: checks if identifier is valid npub

- Updated /e/[id]/page.tsx to use shared parsing utilities
- Updated /p/[id]/page.tsx to use shared parsing utilities
- Reduced code duplication and improved maintainability
- All linting and type checking passes
